New option and remove spaces from buffer names.
Added option g:TagmaBufMgrLastLine to display the Manager Window as the
bottom window with no status line.

Remove spaces from buffer names so formatting is properly applied when
the list of buffers has to wrap to a new line. This is to prevent
buffers with spaces in their names from being split across two lines.